Use generator for `updateChangemaker`
This function is not actually used anywhere, but we do expect it to be
used in the near future so I'm not removing it.
